Move class to `modules/` directory
